Conflicts:
	common/buildcraft/BuildCraftCore.java
	common/buildcraft/BuildCraftFactory.java
	common/buildcraft/builders/BlockArchitect.java
	common/buildcraft/builders/BlockBuilder.java
	common/buildcraft/builders/BlockFiller.java
	common/buildcraft/builders/BlockQuarry.java
	common/buildcraft/builders/BuilderProxyClient.java
	common/buildcraft/builders/TileArchitect.java
	common/buildcraft/builders/TileFiller.java
	common/buildcraft/builders/TileQuarry.java
	common/buildcraft/core/BlockHatched.java
	common/buildcraft/core/lib/render/RenderEntityBlock.java
	common/buildcraft/core/render/RenderBoxProvider.java
	common/buildcraft/factory/BlockFloodGate.java
	common/buildcraft/factory/BlockMiningWell.java
	common/buildcraft/factory/BlockPump.java
	common/buildcraft/factory/FactoryProxy.java
	common/buildcraft/factory/FactoryProxyClient.java
	common/buildcraft/factory/TileMiningWell.java
	common/buildcraft/factory/TilePump.java
	common/buildcraft/transport/TransportGuiHandler.java
	common/buildcraft/transport/gui/ContainerGateInterface.java

+++ common/buildcraft/core/lib/render/RenderEntityBlock.java
@@ -54,7 +54,7 @@ public final class RenderEntityBlock extends Render {
         public TextureAtlasSprite texture = null;
         public TextureAtlasSprite[] textureArray = null;
         public boolean[] renderSide = new boolean[6];
-        public float light = -1f;
+		public int light = -1;
         public int brightness = -1;
 
         public RenderInfo() {
